Add the homepage implementation
The homepage implementation was added based on R3 requirements.
I tried to follow the requirements very closely.

There were a few other things that had to be added that weren't
directly related to the homepage:

- I discovered a bug: if you keep a webpage logged in, shut down
the server, remove the database file, and restart the server,
and then refresh the page, it will cause an error in the server.
I fixed this in frontend.py by invalidating the 'logged_in' key in
the session and redirecting to /login

- I had to add a model for Ticket. It contains a unique ID, name,
quantity, price, and a foreign key relationship to the User who owns it.
Testing was done by manually adding a ticket to the SQLite database:
`insert into ticket (name, quantity, price, owner_id) values ('Test Movie', 2, 1500, 1);`

- I had to reset the database in order to properly add the changes to
the model. In the future I think we may want to look into automating
the creation of the test database (independent of the qa327_test stuff).