Implement R3.5.2: hiding expired tickets
This change filters out expired tickets in the frontend, in order to
avoid showing any expired tickets to the user.
An expiry date has been added to the ticket model to facilitate this.

There are now two tickets in the database:
one expiring on Halloween that does not display on the homepage, and
one expiring on Christmas that does display on the homepage.