✅ Add tests for R3.1 and R3.2
This is mostly just to demonstrate the new common login decorator I made, which should simplify any tests which require logging in as the test user.

There is a warning coming from the frontend, due to a third-party library using a deprecated function. To my understanding, this shouldn't block the test from being put in.