🚀🔬💚 Update tests to use temporary database file
Now the tests will use a temporary database file, instead of having to
potentially remove data from the database at the start of the testcase.

The primary motivation for this was to avoid having to do this only for
the integration tests, which should be using a clean DB imo.

In addition, moving from the `module` scope to `session` scope on the
server fixture gives me excellent performance improvements in the tests,
meaning less CI minutes used.

Some minor refactoring was done to `__init__.py` and `models.py` in
order to use a new DB file and programmatically add the correct tables.