🐛 Fix bug relating to SQL session
There were multiple failures on different systems due to a SQL session
remaining open. It turns out that executing raw SQL does not commit the
changes to the DB file, resulting in the file staying open indefinitely.

One line added to commit the changes to the database after clearing the
ticket and user tables between tests.